requirements.txt
behave==1.2.6
selenium==4.17.2
webdriver-manager==4.0.1
python-dotenv==1.0.1
requests==2.31.0
pytest==7.4.0
python-dotenv==1.0.1